Add Poetry to OscilloscoPy
