+++ src/oscilloscopy/__init__.py
@@ -1,0 +1,6 @@
+import usbtmc
+
+from .channel import Channel
+from .generator import Generator
+from .oscilloscope import Oscilloscope
+from .oscilloscopy import list_devices, get_instrument
